Miscellaneous Updates
Fixed build.gradle to deploy to my local vagrant image. Also updated
the code structure. The API interface is implemented from the code of
the Building Block. I was also creating a secondary identical interface
inside the B2 and implementing that with the same code. Updated my code
to call the API interface and deleted the extraneous interface.